requirements.txt
python-telegram-bot==22.12.0
yt-dlp==2025.01.01
